Update Embedding, Update the method for obtaining the API key

+++ src/main/kotlin/org/protogalaxy/fractalfathom/cli/analysis/ir/Embedding.kt
@@ -1,5 +1,5 @@
 package org.protogalaxy.fractalfathom.cli.analysis.ir
 
 data class Embedding(
-    val values: List<Float> // 嵌入向量
+    val values: List<Double> // 嵌入向量
 )
